Added front end and database support for multiple tags.
