+++ web-client/.env.sample
@@ -1,4 +1,4 @@
-# live heroku/render site, for example
+xxx# live heroku/render site, for example
 REACT_APP_SERVER=<your-backend-url-here>/api/v1
 
 REACT_APP_API_SERVER=<your-backend-url-here>/api/v2
